debug: Just yield more after the first few atoms
This is still just sending the first few atoms normally so we can debug
that our async control flow is working properly, but instead of breaking
at that point, just start taking a lot of naps to let other stuff run.

This means we'll still send data. Our frames will be very late, but
we'll remove this again later.